v2/backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-dotenv==1.1.0
google-api-python-client==2.171.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
pydantic>=2.10,<2.11
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
httpx==0.28.1
